requirements.txt
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
langchain_huggingface
